QuickFit3/qfe_spectraviewer:   - added combobox with common laser lines   - added edits to set plot range (wavelength)
QuickFit3/number_and_brightness:
  - bugfix: sometimes the acquisition background file was loaded, not the overview!

QuickFit3: several Linux code fixes

git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@2593 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/fitfunction_general/qffitfunctiongeneral2gaussianvar.cpp
@@ -64,7 +64,7 @@ bool QFFitFunctionGeneral2GaussianVar::estimateInitial(double *params, const dou
         double pW2=0;
         double pH2=0;
         double pP2=0;
-        if (statistics2PeakFind(pP, pW, pP2, pW2, dataX, dataY, N, 0.0, NAN, &pB, &pH, &pH2)) {
+        if (statistics2PeakFind(pP, pW, pP2, pW2, dataX, dataY, N, 0.0, (double)NAN, &pB, &pH, &pH2)) {
             params[PARAM_OFFSET]=pB;
             params[PARAM_AMPLITUDE]=pH;
             params[PARAM_POSITION]=pP;
