+++ plugins/fitfunction_general/qffitfunctiongeneralgaussianvar.cpp
@@ -58,7 +58,7 @@ bool QFFitFunctionGeneralGaussianVar::estimateInitial(double *params, const doub
         double pW=0;
         double pB=0;
         double pH=0;
-        double pP=statisticsPeakFind(pW, dataX, dataY, N, 0.0, NAN, &pB, &pH);
+        double pP=statisticsPeakFind(pW, dataX, dataY, N, 0.0, double(NAN), &pB, &pH);
         if (statisticsFloatIsOK(pP)) {
             params[PARAM_OFFSET]=pB;
             params[PARAM_AMPLITUDE]=pH;
