+++ plugins/fitfunction_general/qffitfunctiongenerallognormal.cpp
@@ -57,7 +57,7 @@ bool QFFitFunctionGeneralLogNormal::estimateInitial(double *params, const double
         double pW=0;
         double pB=0;
         double pH=0;
-        double pP=statisticsPeakFind(pW, dX.data(), dataY, N, 0.0, NAN, &pB, &pH);
+        double pP=statisticsPeakFind(pW, dX.data(), dataY, N, 0.0, (double)NAN, &pB, &pH);
         if (statisticsFloatIsOK(pP)) {
             params[PARAM_OFFSET]=pB;
             params[PARAM_AMPLITUDE]=pH;
